Changed daily total revenue api to point to proper table.

Related work items: #6202

+++ ABMVantage.Data/Models/Revenue.cs
@@ -8,6 +8,6 @@ namespace ABMVantage.Data.Models
 {
     public class Revenue:ModelBase
     {
-        public double TotalRevenue { get; set; }
+        public decimal TotalRevenue { get; set; }
     }
 }
